Implement touch input handling for rotation and movement
Added touch input handling to the InputManager class introducing functionality for detecting single touch for rotation and continuous touch for movement.
A new event called "Rotate" is defined to notify listeners when a tap input is detected.
A new event called "Move" is defined, which passes a DragDirection enum to notify listeners about the direction of movement input.
A new method, EnablePlayerInput(bool enable), is introduced to enable or disable player input externally.
The method IsTouchOverUI(Vector2 touchPosition) is added to determine if the touch position is over UI elements.

Note to future: was not able to pass the position an have a tap interaction simultanouslly through unity input system. So I am doing so through Touchscreen.current.primaryTouch.position.ReadValue(). I need to check if thats ok.