Added working score to grid
When I grid initialize it will signup it for the game in the game manager. The game manager have a dictionary with the grids ids and its scores. The Running game manager state will receive the game scores from the gridscoremanager scrip that makes the grid score calculations and update the dictionary acordinly.